Move `ISamplePlaybackDisabler` to more general namespace
